Our project is about chicken wings.
This project will be a restaurant about wings, containing a menu with various flavors. Also making a website that contains what we want to be in it. 
Numerous variants of chicken wings and a bunch of add-ons to the project is want we want to make the website look good.